Task/add immutable calculate for behaviors

+++ Tests/fixtures/entity/VehicleTest.cs
@@ -7,7 +7,7 @@ namespace Tests.fixtures.entity
     public class VehicleTest : MovingEntity
     {
         public const int DefaultHeight = 10, DefaultWidth = 10;
-        private const int DefaultMass = 30, DefaultMaxSpeed = 150;
+        private const float DefaultMass = 30, DefaultMaxSpeed = 150;
         public VehicleTest(Vector position, IWorld world) : base(position, world, DefaultHeight, DefaultWidth)
         {
             Velocity = new Vector(0, 0);
